.github/actions/generate_ansible_collection/requirements.txt
ruamel.yaml
pbr